requirements.txt
notion_client==2.0.0
pandas==1.5.2
praw==7.7.0
python-dotenv==1.0.0
python_dateutil==2.8.2
requests==2.28.2
tweepy==4.13.0
playwright==1.32.1
pyperclip==1.8.2
clean-py==0.5
langchain==0.0.157
openai==0.27.6
argilla==1.5.1